pyproject.toml
[tool.poetry]
name = "guardepi"
version = "0.1.0"
description = "Gerenciamento de EPI"
authors = ["Robotz213 <nicholas@robotz.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-mysqldb = "^2.0.0"
flask-mail = "^0.10.0"
python-dotenv = "^1.0.1"
flask-wtf = "^1.2.2"
flask-login = "^0.6.3"
flask-sqlalchemy = "^3.1.1"
openpyxl = "^3.1.5"
pypdf = "^5.1.0"
reportlab = "^4.2.5"
bcrypt = "^4.2.0"
babel = "^2.16.0"
pytz = "^2024.2"
requests = "^2.32.3"
flask-talisman = "^1.1.0"
pillow = "^11.0.0"
pandas = "^2.2.3"
deep-translator = "^1.11.4"
psycopg = "^3.2.3"
eventlet = "^0.38.0"
clear = "^2.0.0"
celery = "^5.4.0"
psycopg2 = "^2.9.10"
redis = "^5.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.10.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black" # Garante compatibilidade com Black
line_length = 88  # O mesmo limite de linha do Black
